change approved project database and create approved project credit source table and mode 1396/7/19

+++ Modules/Budget/Database/Migrations/2017_09_25_075430_create_capital_assets_projects_table.php
@@ -13,46 +13,36 @@ class CreateCapitalAssetsProjectsTable extends Migration
      */
     public function up()
     {
-        Schema::create('tbl_capital_assets_projects', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('cpUId')->length(10)->unsigned();
-            $table->integer('cpCapId')->length(10)->unsigned();
-            $table->integer('cpCoId')->length(10)->unsigned();
-            $table->integer('cpTsId')->length(10)->unsigned();
-            $table->integer('cpHtrId')->length(10)->unsigned();
-            $table->string('cpSubject');
-            $table->string('cpCode');
-            $table->string('cpStartYear');
-            $table->string('cpEndOfYear');
-            $table->tinyInteger('cpPhysicalProgress')->default(0);
-            $table->longText('cpDescription')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('tbl_capital_assets_projects')) {
+            Schema::create('tbl_capital_assets_projects', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('cpUId')->length(10)->unsigned();
+                $table->integer('cpCapId')->length(10)->unsigned();
+                $table->integer('cpCoId')->length(10)->unsigned()->nullable();
+                $table->string('cpSubject');
+                $table->string('cpCode');
+                $table->string('cpStartYear');
+                $table->string('cpEndOfYear');
+                $table->tinyInteger('cpPhysicalProgress')->default(0);
+                $table->longText('cpDescription')->nullable();
+                $table->timestamps();
 
-            $table->foreign('cpUId')
-                ->references('id')->on('users')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
+                $table->foreign('cpUId')
+                    ->references('id')->on('users')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
 
-            $table->foreign('cpCapId')
-                ->references('id')->on('tbl_capital_assets_approved_plan')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
+                $table->foreign('cpCapId')
+                    ->references('id')->on('tbl_capital_assets_approved_plan')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
 
-            $table->foreign('cpCoId')
-                ->references('id')->on('tbl_counties')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
-
-            $table->foreign('cpTsId')
-                ->references('id')->on('tbl_capital_assets_tiny_seasons')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
-
-            $table->foreign('cpHtrId')
-                ->references('id')->on('tbl_how_to_run')
-                ->onDelete('restrict')
-                ->onUpdate('cascade');
-        });
+                $table->foreign('cpCoId')
+                    ->references('id')->on('tbl_counties')
+                    ->onDelete('restrict')
+                    ->onUpdate('cascade');
+            });
+        }
     }
 
     /**
@@ -62,6 +52,8 @@ class CreateCapitalAssetsProjectsTable extends Migration
      */
     public function down()
     {
+        DB::statement('SET FOREIGN_KEY_CHECKS = 0');
         Schema::dropIfExists('tbl_capital_assets_projects');
+        DB::statement('SET FOREIGN_KEY_CHECKS = 1');
     }
 }
